+++ NEWS.md
@@ -1,2 +1,3 @@
 # criticalpath 0.1.0
 
+First version published.